feat(DTFS2-7578): separate out clone functionality into services
